==> 03-multicollinearity.Rmd <==
---
output:
  pdf_document: default
  html_document: default
---
# Checking for Multicollinearity 

_**Multicollinearity  is defined as linear or near linear dependence between predictors. This can impact the usefulness of the linear regression model because it reduces or undermines the statistical significance of the predictor.**_

```{r}
library(faraway)

data1 <- read.csv("cleaned_data_2.csv", fileEncoding="UTF-8-BOM")
data <- read.csv("cleaned_data_scaled_2.csv", fileEncoding="UTF-8-BOM")

```

From the descriptive analysis (pairs plot), we saw that the predictors Length1, Length2, and Length3 were all highly correlated. Height and Width also looked highly correlated with with he other predictors as well, but formed groups of linear relationships. We can quantify the multicollinearty present in the model using the VIF (Variance Inflation Factor). But first, lets view a summary of our initial model.

```{r}


model <- lm(Weight ~ . - 1, data = data)
summary(model)

```

Notice that the regression coefficients for Length1 is negative. This goes against our intuition, since it seems reasonable to think that the greater the length of the fish, the more it is going to weigh (regardless of what is actually being measured). Also note that these predictors do not have great significance. These issues both strongly indicate that multicollinearity is present.

```{r}

vif(model)

```

The VIF for the predictors shows that Length1, Length2, and Length3 are a major concern. They are almost completely linearly dependent. This makes sense given that these measurements are most likely measuring the length of the fish in three, almost identical ways. For this reason, we are justified in only keeping one of the lengths. 

```{r}

data$Length2 = NULL
data$Length3 = NULL
model <- lm(Weight ~ . - 1, data = data)
summary(model)

vif(model)

```

We can see that removing these predictors has fixed the major issues multicollinearity was causing. The estimated regression coefficient for Length1 is now positive and has a significant linear relationship with the response. Note that Length1 and Width still have high VIF (Large VIF). This is likely due to some interaction with the indicator variables. Since Length1 and Width appear to be key variables, they can be left in for now. Hopefully after the model selection process is complete, the VIF's are more reasonable.


```{r}

data1$Length2 <- NULL
data1$Length3 <- NULL

write.csv(data1, 'cleaned_data_3.csv', row.names = FALSE)
write.csv(data, 'cleaned_data_scaled_3.csv', row.names = FALSE)
```
